VIC (Virtual Institutional Counselor)/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
pandas==2.2.3
numpy==1.26.4
simplejson==3.18.4
xmltodict==0.13.0
pathlib==1.0.1
requests==2.32.3
urllib3==2.2.2
tqdm==4.66.4
python-dotenv==1.0.1
gunicorn==21.2.0
bcrypt==4.2.1
PyJWT==2.10.0
pypdf==4.2.0
langchain==0.3.8
transformers==4.45.2
sentence-transformers==2.2.2
huggingface-hub==0.23.2
langchain-ollama==0.2.0
#streamlit==1.40.2
#streamlit-chat==0.1.1
#PyPDF2==3.0.1

# CPU-only PyTorch
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.1+cpu
torchvision==0.17.1+cpu